requirements.txt
flask==2.3.3
requests==2.31.0
python-dotenv==1.0.0

# Uncomment below for Azure AI Services implementation
# azure-ai-textanalytics==5.2.0